reorganize fingerprint utils
